+++ src/components/adminservice/detail/adminserviceDetailNavbar.tsx
@@ -11,6 +11,7 @@ export function AdminserviceDetailNavbar({
   id
 }: AdminserviceDetailNavbarProps): JSX.Element {
   const router = useRouter();
+  console.log("id",id)
   return (
     <div className="flex flex-row items-center justify-between bg-white sticky top-0 h-20 px-10 py-5 min-w-[1200px] border-b  border-gray-300 z-10">
       <div
@@ -24,17 +25,13 @@ export function AdminserviceDetailNavbar({
         </div>
       </div>
       <div className="h-full flex flex-row items-center gap-6 relative">
-        {/* <button
-    className="bg-white text-defaultColor text-base h-full px-7 flex items-center gap-3 rounded-lg w-32 text-center justify-center border border-defaultColor"
-    onClick={() => router.push("/adminservice")}
-    type="button"
-  >
-    ยกเลิก
-  </button> */}
         <button
           className="bg-defaultColor text-white text-base h-full px-7 flex items-center gap-3 rounded-lg w-32 text-center justify-center"
-          onClick={() => router.push(`/adminservice/edit/${id}`)}
-        >
+          onClick={() => {
+            console.log("id: onclick for navigate", id); // ตรวจสอบค่า id
+            router.push(`/adminservice/edit/${id}`);
+          }}
+          >
           แก้ไข
         </button>
       </div>
